Feature/fix gcc warnings

+++ impl/jamtemplate/common/vector_factory.cpp
@@ -7,5 +7,5 @@ jt::Vector2f jt::VectorFactory::fromCartesian(float x, float y) { return jt::Vec
 jt::Vector2f jt::VectorFactory::fromPolar(float r, float phi_in_deg)
 {
     float const phi_in_rad = jt::MathHelper::deg2rad(phi_in_deg);
-    return jt::Vector2f { r * cos(phi_in_rad), r * sin(phi_in_rad) };
+    return jt::Vector2f { r * static_cast<float>(cos(phi_in_rad)), r * static_cast<float>(sin(phi_in_rad)) };
 }
